backend-app/requirements.txt
fastapi>=0.68.0
uvicorn[standard]>=0.15.0
pydantic>=1.8.0,<2.0.0
python-dotenv>=0.19.0
python-multipart>=0.0.5

# Dependencias del agente
langchain>=0.1.0
langgraph>=0.0.15
langchain-openai>=0.0.1
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0

# Utilidades
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5

# Desarrollo
autopep8>=2.0.0
pylint>=2.12.2
pytest>=6.2.5
pytest-asyncio>=0.15.1
